+++ clients/admin-ui/src/types/api/models/DatastoreStagedResource.ts
@@ -33,15 +33,4 @@ export type DatastoreStagedResource = {
   data_uses?: Array<string> | null;
   source_modified?: string | null;
   classifications?: Array<Classification>;
-  /**
-   *
-   * Represents the presence of various diff statuses of the staged resource's children. This is calculated:
-   * - during detection
-   * - upon queuing classification
-   * - after classification finishes
-   * - upon queuing promotion
-   * - after promotion finishes
-   *
-   */
-  child_diff_statuses?: Record<string, boolean>;
 };